+++ src/zinc/defaults.py
@@ -8,11 +8,7 @@ This module provides the Zinc configuration defaults.
 
 """
 
-
-#ZINC_FORMAT = "1"
-
 defaults = dict()
-
 defaults['zinc_format'] = '1'
 defaults['catalog_index_name'] = 'index.json'
 defaults['catalog_config_name'] = 'config.json'
