+++ src/zinc/errors.py
@@ -1,0 +1,14 @@
+
+### Errors ###################################################################
+
+class ZincError(object):
+    def __init__(self, code, message):
+        self.code = code
+        self.message = message
+
+# TODO: might be worst python ever
+class ZincErrors(object):
+    OK = ZincError(0, "OK")
+    INCORRECT_SHA = ZincError(1, "SHA did not match")
+    DOES_NOT_EXIST = ZincError(2, "Does not exist")
+
