+++ src/zinc/storages/aws.py
@@ -1,0 +1,4 @@
+from zinc.catalog import StorageBackend
+
+class AWSStorageBackend(StorageBackend):
+    pass
