+++ src/zinc/tasks/bundle_clone.py
@@ -2,9 +2,8 @@ import os
 import logging
 from shutil import copyfile
 
-from zinc.utils import makedirs
-from zinc.models import bundle_id_for_catalog_id_and_bundle_name, bundle_descriptor_for_bundle_id_and_version
-from zinc.utils import gunzip
+from zinc.utils import *
+from zinc.helpers import *
 
 class ZincBundleCloneTask(object):
