+++ src/zinc/utils.py
@@ -54,4 +54,3 @@ def gunzip(src_path, dst_path):
     f_out.close()
     f_in.close()
 
-
